requirements.txt
playwright
beautifulsoup4
markdownify
numpy
tqdm
faiss-cpu
transformers
sentence-transformers
torch 